requirements.txt
# Smart Grocery Tracker - DETR Dependencies
# Vision-Based Automatic Grocery Tracking System

# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face (for pretrained DETR)
transformers>=4.30.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.0.0
albumentations>=1.3.0

# Data handling
pandas>=2.0.0
pycocotools>=2.0.6

# Configuration
pyyaml>=6.0

# Progress bar
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0

# Machine learning utilities
scikit-learn>=1.2.0

# Development (optional)
# jupyter>=1.0.0
# tensorboard>=2.13.0
